Renames ColorPalette to Colors, and associated [light/dark]ColorPalette functions
This increases consistency with Typography and Shapes, and removes confusion over ColorPalette being a generic 'palette' of colors, when in reality it is a specific implementation of the Material color system.

Bug: b/161812111
Test: ./gradlew updateApi
Relnote: "Renamed ColorPalette to Colors, to better map to the Material color system and remove confusion over ColorPalette being a 'generic' theming object, as opposed to being a specific implementation of the Material color system. Additionally renames lightColorPalette and darkColorPalette to lightColors and darkColors respectively."
Change-Id: I9e97690f51cb9eb0e51ff1a57d08f1911f179232