MediaRouter: Deliver group route descriptor to client.
When dynamic grouping is used, onRouteSelected() is
expected to be called with the new group route.

MediaRouter created a RouteInfo instance by converting
MediaRouteDescriptor -> RoutingSessionInfo -> RouteInfo.
In the converting process some information (e.g. extras) was lost.

This CL notifies MediaRouteDescriptor directly to MediaRouter.

In addition, it calls notifySessionCreated when the group info is updated
so onRouteSelected can be called with correct group route info.

Bug: 161831143
Bug: 159954757

Test: manually w/ support v7 demos
Change-Id: Ib8c84dba54add7442186660fe33ad68daa33bfe1