Adds TabConstants object, removes TabRow object, other TabRow polishing
Default values / implementations used by Tab and TabRow now exist in TabConstants, consistent with Buttons and other components. TabPosition has been moved to a top level class, other members of TabRow have moved to TabConstants.

Additionally updates IndicatorContainer() to just be a modifier that can be applied to an indicator, and renames `indicatorContainer` to `indicator` accordingly.

Bug: b/161809544
Test: TabTest
Relnote: "The `TabRow` object has been removed, replacing it with TabConstants. TabRow.TabPosition has moved to be top level (TabPosition), and indicatorContainer has been renamed to `indicator`. See the samples and documentation for detailed information on how to use the updated API, and defaults."
Change-Id: I54d452697a70b9653812e5366420563a3aa19209