Remove marked as deprecated fillMax* modifiers from LazyItemScope
The were added with an intention to guide users to use the correct new fillParentMax* modifiers available in the scope but this solution seems to be giving move cons then pros.
For example the code like this
LazyColumnFor(items) { item ->
    Row(Modifiger.height(100.dp)) {
       Box(Modifier.fillMaxHeight())
    }
}
Will not work correctly as Modifier.fillMaxHeight() on a Box will be marked as deprecated. This box is technically inside LazyItemScope, but is not a direct child of LazyColumnFor so we shouldn't restrict usages of Modifier.fillMaxHeight(). As Kotlin can't understand which composable is the same layout layer and which not probably it is safer to remove this shadowed modifiers.
Similar issue is that the scope is reused by LazyColumn and LazyRow and having
LazyColumnFor(items) { item ->
    Box(Modifier.fillMaxWidth())
}
Box with Box(Modifier.fillMaxWidth()) inside a vertically scrollable layout will be marked as deprecated, but in fact we can use regular modifiers for cross-axis.
Maybe later we can think of another solution to guide users to use fillParentMax* instead of fillMax* for direct children of LazyColumnFor with Lint?

Relnote: Remove marked as deprecated fillMax* modifiers from LazyItemScope as they are making it harder to add such modifiers correctly for items which are not direct children of LazyColumnFor
Fixes: 162501843
Test: manually
Change-Id: Ifa78d7d5956e7f1d903c03aac4fa34b8bef5c425