Fix crash when text field inside Popup
Removed remembering the text field value passed to the BaseTextField. Instead, it just passes the value downstream to the CoreTextField.
Removed text field value state from the primitive overrides of material text fields. Instead, we construct the TextFieldValue object using the string parameter passed to the text field, as well as selection and composition values that are kept remembered inside the text field.

Fixes: 162258732
Test: manually tested popup (using performTextInput() API that allows to send a text to the text field does not allow to repro the bug in tests).
Test: new test; text field tests passed
Change-Id: I4e825ef872518279272c90199b69e70fa23bceb8