Parameterize tests to use both old and new state manager
Parameterize a number of test suites affected
by the usage of FragmentStateManager so that they
run both under the old state manager and the new
system to avoid regressions during the cross over
time period.

Changed the focusedView() test to not run on the
UI thread, fixing flaky issues when running as part
of a parameterized test suite.

Changed the testTimedPostponeEnterPostponedCalledWithZero()
test to wait for the transition to finish before
finishing the test, ensuring that you don't
enable/disable the state manager while a
transition is running.

Test: tests pass
BUG: 139536619
Change-Id: I1e7bd6f5b135345a9a60c95fe5127b00ff3e69df