Change fling behaviour in Modifier.swipeable
This changes what Modifier.swipeable does when the user stops swiping.
Previously, it would use the final velocity with an exponential decay
to calculate a target and animate to one of its neighbouring anchors
based on the positional thresholds. This was problematic, because it
allowed users to essentially 'skip' a state by flinging hard enough.
To fix this, we now always animate to one of the two anchors bounding
the position where the user let go. Furthermore, this also introduces
a velocity threshold parameter (which defaults to 1/8 dp/ms). When the
final velocity of the swipe exceeds that threshold, we will ignore the
positional thresholds and animate to the state we were swiping towards.

Bug: 161995298
Test: Ran DrawerTest, ScaffoldTest and SwitchTest
Relnote: "Optional param velocityThreshold added to Modifier.swipeable."
Change-Id: I698ba84bbb83ba262b8a55d9f70da4bff907eba9