Stop filtering empty pages on insertSeparators
Turns originalPageOffset in TransformablePage into a set of dependent
indices rather than a single Int, and turns originalIndices to be
relative to the min pageOffset in originalPageOffsets as a convention.
In case of drops, PREPEND / APPEND is converted to either the min or max
originalPageOffset, and all dependent pages on that offset is dropped on
the presenter side (while the fetcher side pre-transform, simply drops
that page). With the current set of operations, these are always
guaranteed to be contiguous from list of presented pages and therefore
does not require DiffUtil for drops.

Primary motiviation for this change is to allow presenter to rely
directly on the pageEvent stream to resend hints, and in the same
process simplifies the separator logic as it no longer requires
remapping the entire event stream to offset for added separators.

Relnote: "insertSeparators no longer filters out empty pages allowing
prefetch distance to be respected by the presenter even in cases where
many empty pages are inserted."
Fixes: 162538908
Test: ./gradlew paging:paging-common:test

Change-Id: I9cff68ad1cc33076254681a3260be3f8ae32b598