Polish Scaffold API
make FabPosition enum to be top-level, remove unnecessary exposed functionality, added drawer-related param and a Modifier

Change-Id: Ic6f7bdd21d469227675b5e59a6b32c592c1ff9e5
Relnote: "bottomBarSize, fabSize and others and not available anymore in ScaffoldState. Use Modifier.onPosition instead on the component you'd like to know the size of. contentColor and Modifier pamateres have been added to Scaffold"
Fixes: 161811485
Fixes: 159979032
Fixes: 161450249
Fixes: 157174382
Test: adjusted accordingly