Consider min constraints in relativePaddingFrom
Before this CL, relativePaddingFrom was not considering min constraints.
Therefore, after the before and after paddings were added, it could
occupy less space and get auto-centered. This CL changes this behavior,
making the modifier occupy the min space and position according to the
specified distance (before if specified, otherwise after).

Fixes: 157581375
Test: AlignmentLineTest
Change-Id: I8fccd9d14005108f69da3fa9101cbf13bf5cdc88