Add hasAwaiters, runWithManualClock and TestUiDispatcher
Added hasAwaiters to ManualFrameClock to see if anything is awaiting a
frame from it. Simply exposes BroadcastFrameClock.hasAwaiters.

Added runWithManualClock to run tests that use a ManualFrameClock.

Added TestUiDispatcher.Main to give easy and safe access to the main
coroutine dispatcher on which composition will happen.

Moved all tests in ScrollableTest to use ManualFrameClock to demo the
new features. (Well, actually the new features are the result of moving
these tests to ManualFrameClock)

Bug: 161247083
Test: ./gradlew compose:foundation:foundation:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.foundation.test.ScrollableTest
Relnote: "Added `ManualFrameClock.hasAwaiters` to see if anything is
awaiting a frame from that clock; `runWithManualClock` as a replacement
for `runBlocking` when running tests that need a ManualFrameClock;
`TestUiDispatcher.Main` that gives easy access to the main UI dispatcher
in your tests.

For example:
```
@Test
fun myTest() = runWithManualClock { clock ->
    // set some compose content
    withContext(TestUiDispatcher.Main) {
        clock.advanceClock(1000L)
    }
    if (clock.hasAwaiters) {
        println("The clock has awaiters")
    } else {
        println("The clock has no more awaiters")
    }
}
```
"

Change-Id: I0a85b019ae5f40f52f2c6b78c08958eb7b8e7485