Use a single set of text selection handle bitmaps
Same as in framework, where a single set of bitmaps is tinted with
`android:tint` on the `<bitmap>` wrapper, we already have tinting
performed in AppCompatDrawableManager. No need to carry two sets
of bitmaps if these are tinted at runtime in any case.

Binary size of minified obfuscated v7 demo:
 - before the change 2,409,701 bytes
 -  after the change 2,398,886 bytes
for the total reduction of about 10KB

Bug: 126209808
Bug: 126204182
Test: ./gradlew appcompat:appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.package=androidx.appcompat.widget
Test: manual inspection of text selection handles visuals in the demo app
Change-Id: Ic291c5e426b29de6a53dc971aa81e524634b647d