Lock the flash mode when image capture pipeline is processing
In ImageCapture, all attemps to change flash mode during picture taken will be saved and only the latest flash mode value will be applied after picture taken is complete.

"Relnote: Fixed the issue that flash on some devices would not turn off. It happens when the flash mode is set to FLASH_MODE_ON to take picture, and is changed to FLASH_MODE_OFF when the flash fired. The symptom is like torch mode enabled."

Bug: 162568128
Test: ./gradlew camera:camera-core:test; ./gradlew camera:camera-camera2:test; ./gradlew camera:camera-core:connectedAndroidTest; ./gradlew camera:camera-camera2:connectedAndroidTest
Change-Id: Ib44514dbad0b8103de615e21717b11d45fb65640