Force PreviewView to use TextureView if extension effect is enabled and there is CaptureProcessor set for Preview
In ProcessingSurface, the pixel format of the output surface is hard-coded as PixelFormat.RGBA_8888. Therefore, ImplementationMode.COMPATIBLE mode should be used to provide a TextureView surface when extension effect is enabled and there is CaptureProcessor set for Preview.

Relnote:"Forced PreviewView to use TextureView if extension effect is enabled and the vendor library implementation needs to do a special process on the output surface."

Bug: 165188770
Test: ./gradlew bOS
Change-Id: I0c3cc867847ec0f350119a8d05ace41c631d04de