Add Composable API for getting the currentNavBackStack entry
Added an API for getting the current back stack entry whenever the
navController changes destinations (likely due to a navigate call or
a popBackStack call). This means that users will always
have access to entry at the top of the back stack (which is the one
currently being shown from navigation's perspective).

RelNote: "Added an API for retrieving the currentNavBackStackEntry
whenever the destination changes"
Test: NavHostControllerTest
Bug: 161472731

Change-Id: I9d15bd561f3882bc34fc511205d6efa232763e7c