Fixed exception thrown when using Image with ContentScale and smaller constraints
Updated PainterModifier's modifyConstraint
methods to consume the minimum of either the
scaled width/height and the maximum dimension
provided in the given constraints.

Some ContentScale types such as FillMaxDimension
(or Crop) will scale according to the maximum
dimension which can be larger than the value
in given constraints.

Content will still draw with the desired scaled
size however, will be clipped to fit within
the constraints

Change-Id: Iccd341d5309c633797d4b1da389f99b0dce9abcb
Fixes: 166762500
Test: Added test to ImageTest