Add XRawType
RawTypes are available in java for backwards compatibility and
are NOT available in Kotlin, which makes it impossible to implement
erasure in KSP.

On ther other hand, erasure is very handy to check if an XType
is a known type and act on it.

This CL replaces XType.erasure with XType.rawType where rawType
IS NOT an instance of XType (similar to XMethodType). Instead,
it allows us to abstract raw infromation w/o relying on the underlying
processor and only has APIs neede by room for this purpose.

Javac implementation relies on erasure while KSP implementation
will use star projection.

Bug: 160322705
Test: existing tests
Change-Id: I742d3d463a45af0937b97b18393f3300eb9fd9fc