[Text Selection] Remove SelectionContainer from the Top.
The SelectionContainer was moved to the top early this year. Before that, if the developer wants to enable the selection, a SelectionContainer needs to be use to wrap the content.

It's nice to enable selection without writing code to add a SelectionContainer since it's at the top. However, it introduces a lot of bugs. For example, the text in a button is automatically selected when the button is in the range of the seleted text. In addition, since LongPressAndDrag Gesture is in SelectionContainer which is at the top, it's hard for other components to detect long press and perform other actions.

After a few meetings, we decided to:
1. Remove SelectionContainer from the Top.
2. Move LongPressAndDrag into Text from SelectionContainer.

Bug: 158459594
Bug: 143413207
Bug: 161184848
Bug: 159743502
Test: Manually.

RelNote: Remove SelectionContainer from the Top to disable default selection and avoid unexpected behavior. One can specifically use a SelectionContainer to wrap the content that needs to be selectable instead.

Change-Id: I8dfe067d1a56ddb95c8fdd7fa2678e8ac43bba6b