Fix invalidate() scheduling when watch system time is changed to the past.
If the watch system time is changed to a past value, the next
scheduled frame draw time becomes much greater than the current
system time (generally they have a difference in the 10s of
milliseconds). This large difference causes nextDrawTimeMillis to
remain significantly larger than currentTimeMillis consistently.
The fix reassigns the value of nextDrawTimeMillis when the system
time is changed.

Test: Presubmit
Change-Id: I1086e8d0acd6eec611674ce21372daaace9bf8ee