Verify type converter factories in DAOs
This CL changes how we list factories.
Instead of relying on context which is not very reliable as it may not
be used, instead, this CL relies on generated code for figuring out
which factories are used.

As there is no clear path to collect them, I've added a metadata field
to class writer that allows stashing such information. Then each DAO
generates a method that returns the list of them and generated
RoomDatabase collects the list from each dao. This also allows Room
database to provide a better error on where each type converter
factory is used.

I've also changed not to provide the map as a public API but instead,
provide a getTypeConverterFactory method in the room database so that
in the future we can change how it is kept internally.

Bug:121067210
Relnote: You can now provide a TypeConverterFactory for TypeConverters
for Room to use when constructing them. This is especially useful if
the TypeConverter needs some dependencies (e.g. a serializer).

Change-Id: I78a04da9faf87a8d4d2388c4681f465b326d7635