Bottom sheet scaffold component
This is a new kind of scaffold component, which implements a persistent
bottom sheet and combines with a top app bar, a floating action button,
a navigation drawer, and snackbar support, in order to construct an app
screen. Persistent bottom sheets display content that complements the
primary content of the screen and remain visible while users interact
with the primary content, making them useful for multitasking.

Bug: 113856787
Bug: 148996320
Test: TODO
Relnote: "Added experimental BottomSheetScaffold component."
Change-Id: Ie02f03288c910fbfa6702a77342eb00917050ba2