Ensure transitions with only shared elements end
In DefaultSpecialEffects controller if there is a fragment that has only
a sharedElementTransition it currently does not have a way to end.

DefaultSpecialEffects should check if the transitionInfo has a
sharedElement when it is looking to start transitions, and when checking
if a transition involves a shared element.

Relnote: "Fragments that only have a shared element transition now
properly complete their special effects meaning they actually move to
their final state"
Test: noSharedElementToSharedElementTest
Bug: 166658128

Change-Id: Iaebc76914b9c24968a2943c18911b183f9dd8a4f