Randomly generate activity result request codes
Currently the ActvityResultRegistry just assigns requestCode in order,
starting with an initial value of 0x00010000 going up to max integer.

If registering and unregistering several callback at a time, say by
adding and removing a fragment that has multiple onActivityResult
request, it is possible to use up the 2147418111 ids, which will crash
as a result since the ids will clash.

Instead of assigning request codes in order, we should randomly generate
them and skip any ids that are currently assigned. This decreases our
chances of clashes and also allows us to reuse any free ids that might
be available.

Test: all exisiting test pass
Bug: 168779518
Change-Id: Idab3745dffb231ee4e31124c1ed91894997685c8