Added setting of the extensions mode into the Camera
This change allows for extensions to be applied without creating a new
instance of the UseCases. The reconfiguring occurs under the hood in
CameraUseCaseAdapter.

Extension.setExtension() will set the Configs on the Camera so that when
UseCases are attached to the camera the additional Configs will be
applied to the UseCase. In addition, when setExtension() is called all
currently attached UseCases will be detached then reattached with the
new extensions Configs. This requires the camera to close before an
extension can be applied.

Bug: b/154044340, b/154939118
Test: camera robolectric & android tests
      manual test of camera-core testapp
Change-Id: I523ba1c9e33ca063a6bf2fca58f966d0473313af