Remove findDefaultImpls from XMethodElement
Kotlin generates a stub class for default methods in interfaces.
Room uses this information to implement the delegating call in DAOs.
Unfortunately, in KSP, such DefaultImpls declaration does not exist,
making this API impossible to implement without creating synthetic
elements. On the other hand, the shape of that default implementation
is fully driven by the method itself hence we don't really need to
find the stub to generate the call, we only need to know if it exists.

Hence, this CL changes XMethodElement.findKotlinDefaultImpl  to
hasKotlinDefaultImpl and removes the only access to
that implementation element which was unnecessary.

Bug: 160322705
Test: existing tests
Change-Id: Icd5b894067695906ee74119f9cbd71d0c80aa0d1